wasm-encoder: move component core type encoding to a separate section.
Similar to the previous commit, this moves component core type encoding to a
`CoreTypeSection` rather than extending `TypeSection` to encode for both
modules and components.

This helps make the distinction between a "type section" (module) and a "core
type section" (component) a little more clear until such type that modules and
components can share a core type section format.
