+++ crates/wasm-encoder/src/component/types.rs
@@ -1,7 +1,174 @@
 use crate::{
     encode_section, ComponentOuterAliasKind, ComponentSection, ComponentSectionId,
-    ComponentTypeRef, Encode, TypeEncoder,
+    ComponentTypeRef, Encode, EntityType, ValType,
 };
+
+/// Represents the type of a core module.
+#[derive(Debug, Clone, Default)]
+pub struct ModuleType {
+    bytes: Vec<u8>,
+    num_added: u32,
+    types_added: u32,
+}
+
+impl ModuleType {
+    /// Creates a new core module type.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Defines an import in this module type.
+    pub fn import(&mut self, module: &str, name: &str, ty: EntityType) -> &mut Self {
+        self.bytes.push(0x00);
+        module.encode(&mut self.bytes);
+        name.encode(&mut self.bytes);
+        ty.encode(&mut self.bytes);
+        self.num_added += 1;
+        self
+    }
+
+    /// Define a type in this module type.
+    ///
+    /// The returned encoder must be used before adding another definition.
+    #[must_use = "the encoder must be used to encode the type"]
+    pub fn ty(&mut self) -> CoreTypeEncoder {
+        self.bytes.push(0x01);
+        self.num_added += 1;
+        self.types_added += 1;
+        CoreTypeEncoder(&mut self.bytes)
+    }
+
+    /// Defines an export in this module type.
+    pub fn export(&mut self, name: &str, ty: EntityType) -> &mut Self {
+        self.bytes.push(0x03);
+        name.encode(&mut self.bytes);
+        ty.encode(&mut self.bytes);
+        self.num_added += 1;
+        self
+    }
+
+    /// Gets the number of types that have been added to this module type.
+    pub fn type_count(&self) -> u32 {
+        self.types_added
+    }
+}
+
+impl Encode for ModuleType {
+    fn encode(&self, sink: &mut Vec<u8>) {
+        sink.push(0x50);
+        self.num_added.encode(sink);
+        sink.extend(&self.bytes);
+    }
+}
+
+/// Used to encode core types.
+#[derive(Debug)]
+pub struct CoreTypeEncoder<'a>(pub(crate) &'a mut Vec<u8>);
+
+impl<'a> CoreTypeEncoder<'a> {
+    /// Define a function type.
+    pub fn function<P, R>(self, params: P, results: R)
+    where
+        P: IntoIterator<Item = ValType>,
+        P::IntoIter: ExactSizeIterator,
+        R: IntoIterator<Item = ValType>,
+        R::IntoIter: ExactSizeIterator,
+    {
+        let params = params.into_iter();
+        let results = results.into_iter();
+
+        self.0.push(0x60);
+        params.len().encode(self.0);
+        self.0.extend(params.map(u8::from));
+        results.len().encode(self.0);
+        self.0.extend(results.map(u8::from));
+    }
+
+    /// Define a module type.
+    pub fn module(self, ty: &ModuleType) {
+        ty.encode(self.0);
+    }
+}
+
+/// An encoder for the core type section of WebAssembly components.
+///
+/// # Example
+///
+/// ```rust
+/// use wasm_encoder::{Component, CoreTypeSection, ModuleType};
+///
+/// let mut types = CoreTypeSection::new();
+///
+/// types.module(&ModuleType::new());
+///
+/// let mut component = Component::new();
+/// component.section(&types);
+///
+/// let bytes = component.finish();
+/// ```
+#[derive(Clone, Debug, Default)]
+pub struct CoreTypeSection {
+    bytes: Vec<u8>,
+    num_added: u32,
+}
+
+impl CoreTypeSection {
+    /// Create a new core type section encoder.
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// The number of types in the section.
+    pub fn len(&self) -> u32 {
+        self.num_added
+    }
+
+    /// Determines if the section is empty.
+    pub fn is_empty(&self) -> bool {
+        self.num_added == 0
+    }
+
+    /// Encode a type into this section.
+    ///
+    /// The returned encoder must be finished before adding another type.
+    #[must_use = "the encoder must be used to encode the type"]
+    pub fn ty(&mut self) -> CoreTypeEncoder<'_> {
+        self.num_added += 1;
+        CoreTypeEncoder(&mut self.bytes)
+    }
+
+    /// Define a function type in this type section.
+    pub fn function<P, R>(&mut self, params: P, results: R) -> &mut Self
+    where
+        P: IntoIterator<Item = ValType>,
+        P::IntoIter: ExactSizeIterator,
+        R: IntoIterator<Item = ValType>,
+        R::IntoIter: ExactSizeIterator,
+    {
+        self.ty().function(params, results);
+        self
+    }
+
+    /// Define a module type in this type section.
+    ///
+    /// Currently this is only used for core type sections in components.
+    pub fn module(&mut self, ty: &ModuleType) -> &mut Self {
+        self.ty().module(ty);
+        self
+    }
+}
+
+impl Encode for CoreTypeSection {
+    fn encode(&self, sink: &mut Vec<u8>) {
+        encode_section(sink, self.num_added, &self.bytes);
+    }
+}
+
+impl ComponentSection for CoreTypeSection {
+    fn id(&self) -> u8 {
+        ComponentSectionId::CoreType.into()
+    }
+}
 
 /// Represents a component type.
 #[derive(Debug, Clone, Default)]
@@ -22,11 +189,11 @@ impl ComponentType {
     ///
     /// The returned encoder must be used before adding another definition.
     #[must_use = "the encoder must be used to encode the type"]
-    pub fn core_type(&mut self) -> TypeEncoder {
+    pub fn core_type(&mut self) -> CoreTypeEncoder {
         self.bytes.push(0x00);
         self.num_added += 1;
         self.core_types_added += 1;
-        TypeEncoder(&mut self.bytes)
+        CoreTypeEncoder(&mut self.bytes)
     }
 
     /// Define a type in this component type.
@@ -120,11 +287,11 @@ impl InstanceType {
     ///
     /// The returned encoder must be used before adding another definition.
     #[must_use = "the encoder must be used to encode the type"]
-    pub fn core_type(&mut self) -> TypeEncoder {
+    pub fn core_type(&mut self) -> CoreTypeEncoder {
         self.bytes.push(0x00);
         self.num_added += 1;
         self.core_types_added += 1;
-        TypeEncoder(&mut self.bytes)
+        CoreTypeEncoder(&mut self.bytes)
     }
 
     /// Define a type in this instance type.
